Organize adaptor files
