+++ src/adaptor/discord/ws.ts
@@ -1,5 +1,5 @@
 import type { Client } from 'discord.js';
-import type { Ping } from '../service/ping.js';
+import type { Ping } from '../../service/ping.js';
 
 export class DiscordWS implements Ping {
   constructor(private readonly client: Client) {}
