+++ src/adaptor/message-proxy.ts
@@ -2,7 +2,7 @@ import type {
   MessageEventProvider,
   MessageUpdateEventProvider
 } from '../runner/index.js';
-import type { RawMessage, Transformer } from './transformer/index.js';
+import type { RawMessage, Transformer } from './transformer.js';
 import type { Client } from 'discord.js';
 
 export type MessageHandler<M> = (message: M) => Promise<void>;
