+++ src/server/index.ts
@@ -39,10 +39,10 @@ import {
   registerAllCommandResponder
 } from '../service/index.js';
 import type { AssetKey } from '../service/party.js';
-import { DiscordMemberStats } from '../adaptor/discord-member-stats.js';
-import { DiscordRoleManager } from '../adaptor/discord-role.js';
-import { DiscordSheriff } from '../adaptor/discord-sheriff.js';
-import { DiscordWS } from '../adaptor/discord-ws.js';
+import { DiscordMemberStats } from '../adaptor/discord/member-stats.js';
+import { DiscordRoleManager } from '../adaptor/discord/role.js';
+import { DiscordSheriff } from '../adaptor/discord/sheriff.js';
+import { DiscordWS } from '../adaptor/discord/ws.js';
 import { EmojiProxy } from '../adaptor/emoji-proxy.js';
 import { GenVersionFetcher } from '../adaptor/version/fetch.js';
 import type { KaereMusicKey } from '../service/kaere.js';
